requirements.txt
pytest~=6.2.2
selenium~=3.141.0
requests~=2.25.1
jsonmodels~=2.4.1
PyYAML==5.4.1